requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
frozenlist==1.8.0
idna==3.11
mac-vendor-lookup==0.1.15
multidict==6.7.1
propcache==0.4.1
python-nmap==0.7.1
scapy==2.7.0
typing_extensions==4.15.0
yarl==1.23.0
